Chore : 타입에 따라 swift파일 분할, 변수명 수정, 중복 내용 간소화
1. 타입별로 파일을 생성해 관리하도록 수정
2. 불필요한 argument label 삭제 및 함수명, 변수명 수정
3. private func getRandomMuscleStat(range: ClosedRange<Int>) 함수삭제
4. range 매직넘버 사용하지 않고, 변수로 관리하도록 수정
